+++ odbc-api/src/buffers/text_row_set.rs
@@ -98,8 +98,11 @@ impl TextRowSet {
 }
 
 unsafe impl RowSetBuffer for TextRowSet {
+    fn bind_type(&self) -> u32 {
+        0 // Specify column wise binding.
+    }
+
     unsafe fn bind_to_cursor(&mut self, cursor: &mut impl Cursor) -> Result<(), Error> {
-        cursor.set_row_bind_type(0)?;
         cursor.set_row_array_size(self.batch_size)?;
         cursor.set_num_rows_fetched(&mut self.num_rows)?;
         for (index, column_buffer) in self.buffers.iter_mut().enumerate() {